DB에서 장소 정보를 가져와 여러개의 마커를 띄울 수 있는 로직 생성 및 임의의 5개 장소 마커 띄우기 구현
